requirements.txt
ebaysdk
dnspython
free-proxy
lxml
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
flask_pymongo==2.3.0
bs4==0.0.1
datefinder==0.7.1
requests==2.25.1
python-dateutil==2.8.1
